Move DisplayRepository from SystemUI to displaylib
This moves all the non-sysui specific logic from DisplayRepository of sysui to displaylib.

Tests are still left in SystemUI for now, just to prove that the behaviour of SysUI's display lib didn't change at all.

All the handling for decorations is still in sysui code, as it depends on some sysui specific classes.

A few minor changes:
- PendingDisplay references have been renamed to the new class in the lib
- ConflatedCallbackFlows are now just callbackFlows.conflate explicitly
- Pairwise has been copied unfortunately, as the sysui util is not in a shared lib. We should move it.
- DisplayEvent is now private (it didn't make sense for it to be public)

Bug: 362719719
Bug: 401305290
Test: DisplayRepositoryTest, builds and runs successfully
Flag: NONE - Just moving code around
Change-Id: I96d4e422926f908633e7006fbc5786dabf372b90