Implement `ViewMotionValue` runtime independent of compose
Move the logic to `Computations` in a mechanic refactor. Neither public
API nor any logic is changed.

The goal of this separate the `Snapshot` related code from the
computations. The existing computation steps are split into 4 separate
interfaces, to ensure the correct order of the computations (this is an
improvement from the current implementation). Every subsequent interface
exposes the computation result as a `val`, allowing to either implement
this with state observation, or - as intended in the next CLs for the
view-based implementation - as sequential calls to `compute*()`

Flag: com.android.systemui.scene_container
Bug: 389637975
Test: atest mechanics_test
Change-Id: I66dbac24925a21bade7a9628012223a8a7694d1d