[revert^2] Move PerDisplayRepository from SystemUI to displaylib
This moved PerDisplayRepository to displaylib, making it available to places outside of sysui.

A thing worth noticing is that the factory is not actually generated from the dagger component of displaylib, as we can't directly expose an element with a generic constructor.

The way this works is that PerDisplayRepository takes part in the dagger graph of systemui (or any other project compiling this).

For this reason, the @DisplayLibBackground annotation has been introduced, to allow systemui to provide a coroutine scope usable from PerDisplayRepositoryImpl.Factory.

Tests are untouched, and still in SystemUI (by purpose, to prove the behavior didn't change)

In a follow up I'll remove the displaylib dagger component (not strictly needed here), and move the tests to displaylib.

Bug: 401305290
Test: PerDisplayInstanceRepositoryImplTest, builds and runs successfully
Flag: NONE - Just moving code around
Change-Id: I6e48af35737a9a33edbfcdcd8809035cf817c6d9