Add support for custom lifecycle manager for PerDisplayRepository
This allows to have custom policies to clean up instances based on other signals (e.g. display having screen decorations)

Tests are kept in sysui (see cl in the topic), mainly to keep the ability to use kosmos

Bug: 362719719
Bug: 401305290
Test: PerDisplayInstanceRepositoryImplTest
Flag: NONE - adding a functionality to the lib, but not using it yet.
Change-Id: Ie3d035decf2192018b12e0aead731fd0ba613d95